_other_scripts/requirements.txt
GitPython==3.1.27
comment-parser==1.2.4